requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
langchain_openai
faiss-cpu
streamlit
pandas
python-dotenv
fastapi 
uvicorn  
groq 
tavily-python
